organized outputs into output package
